♻️Refactor(GATHERING): GatheringRepositoryCustomImpl - radiusDistanceCondition 수정
- 파라미터 값 Double 에서 BigDecimal 로 수정
- Expression<BigDecimal> 로 수정
- 잘못 된 주석 수정(위도 와 경도를 반대로 적었었음)

related: #103, #33

+++ src/main/java/com/runto/domain/gathering/dao/GatheringRepositoryCustomImpl.java
@@ -19,6 +19,7 @@ import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StringUtils;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -276,15 +277,14 @@ public class GatheringRepositoryCustomImpl implements GatheringRepositoryCustom 
         return gathering.title.contains(searchTitle);
     }
 
-    
-    // TODO: 계산 쿼리 수정필요
-    private BooleanExpression radiusDistanceCondition(Double radiusDistance, Double x, Double y) {
+
+    private BooleanExpression radiusDistanceCondition(Double radiusDistance, BigDecimal x, BigDecimal y) {
 
         if (isGeoRadiusValid(radiusDistance, x, y)) {
 
             // 좌표를 Expression<Double>로 변환
-            Expression<Double> latitude = create(y); // 경도(longitude)는 x 좌표
-            Expression<Double> longitude = create(x); // 위도(latitude)는 y 좌표
+            Expression<BigDecimal> latitude = create(y);  // 위도(latitude)는 y 좌표
+            Expression<BigDecimal> longitude = create(x); // 경도(longitude)는 x 좌표
 
 
             // 지구 반지름: km 단위로 설정 (6371km)
@@ -323,7 +323,7 @@ public class GatheringRepositoryCustomImpl implements GatheringRepositoryCustom 
         return null;
     }
 
-    private boolean isGeoRadiusValid(Double radiusDistance, Double x, Double y) {
+    private boolean isGeoRadiusValid(Double radiusDistance, BigDecimal x, BigDecimal y) {
         return radiusDistance != null && (x != null && y != null);
     }
